Add supplier profile strength endpoint
