+++ src/client/components/SupplierProfileStrength/SupplierProfileStrength.react.js
@@ -1,43 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import Gauge from 'react-svg-gauge';
-
-
-function hue2rgb(p, q, t){
-  if(t < 0) t += 1;
-  if(t > 1) t -= 1;
-  if(t < 1/6) return p + (q - p) * 6 * t;
-  if(t < 1/2) return q;
-  if(t < 2/3) return p + (q - p) * (2/3 - t) * 6;
-  return p;
-}
-
-function hslToRgb(h, s, l){
-  var r, g, b;
-
-  if(s == 0) {
-    r = g = b = l; // achromatic
-  } else {
-      var q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-      var p = 2 * l - q;
-      r = hue2rgb(p, q, h + 1/3);
-      g = hue2rgb(p, q, h);
-      b = hue2rgb(p, q, h - 1/3);
-  }
-
-  return [Math.floor(r * 255), Math.floor(g * 255), Math.floor(b * 255)];
-}
-
-function getHexColor(value) {
-  var string = value.toString(16);
-  return (string.length === 1) ? '0' + string : string;
-}
-
-function getHslColor(i) {
-  var hue = i * 1.2 / 360;
-  var rgb = hslToRgb(hue, 1, 0.4);
-
-  return '#' + getHexColor(rgb[0])  + getHexColor(rgb[1])  + getHexColor(rgb[2]);
-}
+import request from 'superagent-bluebird-promise';
+import hexColourCalculator from '../../utils/hexColourCalculator.js';
 
 class SupplierProfileStrength extends Component {
   static propTypes = {
@@ -48,12 +12,27 @@ class SupplierProfileStrength extends Component {
   constructor() {
     super();
     this.state = {
-      value: 45
+      value: 0
+    }
+  }
+
+  profileStrengthPromise = null;
+
+  componentDidMount() {
+    const url = `${this.props.actionUrl}/supplier/api/suppliers/${encodeURIComponent(this.props.supplierId)}/profile_strength`;
+    this.profileStrengthPromise = request.get(url).set('Accept', 'application/json').promise();
+
+    this.profileStrengthPromise.then(response => this.setState({ value: response.body })).catch(errors => null);
+  }
+
+  componentWillUnmount() {
+    if (this.profileStrengthPromise) {
+      this.profileStrengthPromise.cancel();
     }
   }
 
   render() {
-    const colorHex = getHslColor(this.state.value);
+    const colorHex = hexColourCalculator.colourFor(this.state.value);
 
     return (
       <div>
