+++ src/server/routes/supplier_profile_strength.js
@@ -1,0 +1,4 @@
+
+module.exports = function(app, db, config) {
+  app.get('/api/suppliers/:supplierId/profile_strength', (req, res) => res.json(46));
+}
